+++ R/RUtreebalance.R
@@ -154,7 +154,7 @@
 #' Calculate tree balance index \ifelse{html}{\out{J<sup>1</sup>}}{\eqn{J^1}}
 #' (when `nonRootDominance = FALSE`) or
 #' \ifelse{html}{\out{J<sup>1c</sup>}}{\eqn{J^{1c}}}
-#' (when `nonRootDominance = TRUE`) from \insertRef{Lemant2022}{TreeTools}.
+#' (when `nonRootDominance = TRUE`) from \insertCite{Lemant2022}{TreeTools}.
 #' 
 #' If population sizes are not provided, then the function assigns size 0 to
 #' internal nodes, and size 1 to leaves.
@@ -173,6 +173,11 @@
 #' \insertCite{Lemant2022;textual}{TreeTools}, page 1223.
 #' @param nonRootDominance Logical specifying whether to use non-root dominance
 #' factor.
+#' 
+#' @returns `J1Index()` returns a numeric specifying the
+#' \ifelse{html}{\out{J<sup>1</sup>}}{\eqn{J^1}} index of `tree`.
+#' \eqn{J^1(T) = 1} for a perfectly balanced tree; 
+#' \eqn{J^1(T) = 0} for a pectinate (linear / caterpillar) tree.
 #' 
 #' @examples
 #' # Using phylo object as input: